Break up BitcoinCoreConnectionSettings.tsx into smaller files
