+++ src/renderer/ipc/ipcService.ts
@@ -1,5 +1,5 @@
 import { PromiseType } from "_t/typeHelpers";
-import { ShowErrorMtM, RpcRequestMtM } from "_t/IpcMessages";
+import { ShowErrorMtM, RpcRequestMtM, RpcConfigurations } from "_t/IpcMessages";
 import { callMain } from "./callMain";
 
 const extractPayload = async <T extends ReturnType<typeof callMain>>(
@@ -43,6 +43,11 @@ class IpcService {
         payload: request,
       }),
     );
+
+  saveBitcoinCoreConnectionConfigurations = (
+    nonce: NONCE,
+    connectionConfigurations: RpcConfigurations | null,
+  ) => {};
 }
 
 export const ipcService = new IpcService();
